refactor events controllers

+++ CMS/Controllers/EventsController.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
+﻿using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using CMS.Models.CMSModel;
 
@@ -13,13 +9,15 @@ namespace CMS.Controllers
     [Authorize(Roles = "SuperAdmin, Administration")]
     public class EventsController : Controller
     {    
-        private CmsContext _db = new CmsContext();
+        private readonly CmsContext _db = new CmsContext();
 
         public async System.Threading.Tasks.Task<ActionResult> Index()
         {
-            var events = GetEventList(); ;
+            var events = GetEventList(); 
             return View(await events.ToListAsync());
         }
+
+        // GET: Events
         public IQueryable<Event> GetEventList()
         {
             IQueryable<Event> events = _db.Events
@@ -29,35 +27,6 @@ namespace CMS.Controllers
             return events;
 
         }
-
-        //  GET: Events by DeviceId for Raspberry Pi
-        //[HttpGet]
-        //[Route("device/{deviceId}/device-events")]
-        //public List<Event> DeviceEventsList(int deviceId)
-        //{
-        //    DateTime now = DateTime.UtcNow.Date;
-        //    var deviceEvents = db.Events
-        //                        .Join(db.Devices,
-        //                            evnt => evnt.LocationId,
-        //                            dev => dev.LocationId,
-        //                            (evnt, dev) => new { Event = evnt, Device = dev })
-        //                        .Where(evntAndDev => evntAndDev.Device.DeviceId == deviceId
-        //                            && evntAndDev.Event.StartTime >= now)
-        //                        .Select(evntAndDev => new { evntAndDev.Event });
-        //    return deviceEvents.ToList();
-        //}
-
-
-        // GET: Events
-        //public ActionResult Index()
-        //{
-        //     var events = _db.Events.Include(ec => ec.AssociatedEvent).Include(ec => ec.AssociatedLocation);
-        //     return View(events.ToList());
-
-        //}
-
-
-
 
         // GET: Events/Details/5
         public ActionResult Details(int? id)
@@ -84,8 +53,6 @@ namespace CMS.Controllers
         }
 
         // POST: Events/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "EventId,Name,Details,Priority,StartTime,EndTime,OrganiserId,EventCategoryId,LocationId")] Event @event)
@@ -122,8 +89,6 @@ namespace CMS.Controllers
         }
 
         // POST: Events/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "EventId,Name,Details,Priority,StartTime,EndTime,OrganiserId,EventCategoryId,LocationId")] Event @event)
@@ -161,7 +126,7 @@ namespace CMS.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Event @event = _db.Events.Find(id);
-            _db.Events.Remove(@event);
+            if (@event != null) _db.Events.Remove(@event);
             _db.SaveChanges();
             return RedirectToAction("Index");
         }
